api/requirements.txt
django==5.0.6
djangorestframework==3.15.1
numpy==1.26.1
pandas==2.2.2
scikit-learn==1.5.0